# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book Store API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/app/books": {
            "get": {
                "tags": [
                    "APP"
                ],
                "summary": "Danh sách Books ",
                "operationId": "app_book_list",
                "parameters": [
                    {
                        "name": "X-localication",
                        "in": "header",
                        "description": "Ngôn ngữ hiển thị",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "vi"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy danh sách sách thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/app/collection/{slug}": {
            "get": {
                "tags": [
                    "APP"
                ],
                "summary": "Lấy danh sách Books theo category slug",
                "operationId": "app_book_listBySlug",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug của danh mục",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "van-hoc"
                        }
                    },
                    {
                        "name": "X-localication",
                        "in": "header",
                        "description": "Ngôn ngữ hiển thị",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "vi"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy danh sách sách thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/app/books/{slug}": {
            "get": {
                "tags": [
                    "APP"
                ],
                "summary": "Chi tiết thông tin của Book",
                "operationId": "app_book_detailBySlug",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug của book",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "van-hoc"
                        }
                    },
                    {
                        "name": "X-localication",
                        "in": "header",
                        "description": "Ngôn ngữ hiển thị",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "vi"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy danh sách sách thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/app/categories/{slug}": {
            "get": {
                "tags": [
                    "APP"
                ],
                "summary": "Lấy chi tiết danh mục cha kèm danh mục con",
                "operationId": "detailCategory",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug của danh mục",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "van-hoc"
                        }
                    },
                    {
                        "name": "X-localication",
                        "in": "header",
                        "description": "Ngôn ngữ hiển thị",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "vi"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy danh sách sách thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}